Added the JavaScript files from the previous project, and added DOM change listeners.
Pasted from the version control document:
In the end we don’t want to make modifications which are tailored to one
Web applications to the usaProxy. Information obtained from the
responses was analyzed, but as they were only returning data, that
doesn’t enable a reconstruction of what the user was doing, instead of
that, I will have to obtain the DOM information and record it. In the
new proxyscript version I removed all iKUP specific functions. I left
the xpath recording for the events, time modification and jQuery
addition but I removed the rest. The function $(document.body).html();
gives us the DOM we are looking for. The usaProxy script records the
events’ information in the variable logVal_UsaProxy. In the java server,
the second case of the function processRequest (ClientRequest.java) gets
an html file from the client via: InputStream fis =
usaProxy.getFileSender().getHTMLData(filename); I should use an
httprequest to send the DOM, but for now I will use another kind of
event which will be called DOMCHANGE. The usaProxy server will recognize
it in the step of processing the event (EventManager.java) and record it
to the correspondent file. I added the following code to the log
function in EventManager.java String sid = HTTPData.getValue(data,
"sid"); String sd = HTTPData.getValue(data, "sd");
…......

//TODO I will have to check what event is being recorded in order to
send it to logDOMChange if (dataArray[m].contains("domchange"))
logDOMChange(out, data, client, sd, sid);
Created a folder called DOMchanges, in which different folders will be
created for each user.