Changed the filenames for storing the DOM changes. Windows cannot have the ":" I was using for the times.
The problem with finding different number of newlines at the end of the
DOMs was still there. I fixed it by removing ALL newlines from the
strings to be compared. This change only affects the compare engine, so
the newlines will remain untouched in the text file.

The proxy now always records three digits for the milliseconds.