Finished creating Java classes for all events.
Created a HashMap to parse the event log data, simplifying the entire
process.

Events are now committed to the MongoDB via the MongoDAO

In order to be able to work with events queried from the database in
Java, I created a GenericEvent class that will contain empty values for
all existent attributes, which will then be overwritten by the data
available for each particular event.