I changed the name of the functions in the JavaScript related to the browser’s information to refer to client’s information. They now return the client’s operating system as well, and add it to the event string with the attribute name of “platform”. Now I have to change the server so all the beans take it into account. A major change in the way events beans handle inheritance took place. The specific events now only contain the information which is unique to them. The GenericEvent gained importance, as it now contains ALL the shared information between the events. Single events will now rely on the super constructor to fill the values for those common values, and then they will fill the values specific to them.
EventDataHashMap now decodes all the values before storing them into the
hashmap. This way all the URLs and string values are readable.

A system to tackle privacy was designed. In the copypastecode an array
with all the ids of the sensitive values will be manually modified by
the user. The JavaScript will then take this array into account in order
to infer if the target element’s event should be recorded.
//Sensitive fields' IDs should be in this array. It will be empty by
default
var protectedIds=["id1","id2","id3"];
	
The function privacyCheck, will see if the provided ID is sensitive, and
will even check if it’s in a node contained in it.
/**
 * This function will take an event as input, and it will compare its
target's parents' id
 * against the ones contained in the "protectedIds" variable. It returns
a boolean
 * indicating if the given target is protected or not
 * 
 * BEWARE!!! The ids comparison is case sensitive!
 */ 
function privacyCheck(ev){
	
	jqueryTarget = $(ev.target);
	var target 	= (isNotOldIE) ? ev.target : ev.srcElement;
	targetId = target.id;
	
	for (i=0;i<protectedIds.length;i++)
	{
		if (targetId == protectedIds[i]){
			console.log("Target has banned ID!! " + protectedIds[i]);
			return true;
		}
		
		//The '#' is required for jQuery to know it's an id
		if (jqueryTarget.parents('#' + protectedIds[i]).length){
			console.log("Target is child of banned ID!! " + protectedIds[i]);
			return true;
		}

    }

	return false;
}


This function will then be called before calling events that could pose
breaches in privacy, using the following code:
if privacyCheck(target.id)
	return true;